+++ api-test/src/main/java/com/cinema/api/ProjectConfig.java
@@ -1,0 +1,5 @@
+package com.cinema.api;
+
+public interface ProjectConfig {
+    //можно сделать юрл ссылку и ссылаться на неё
+}
