+++ api-test/src/main/java/com/cinema/api/conditions/Condition.java
@@ -1,0 +1,8 @@
+package com.cinema.api.conditions;
+
+import io.restassured.response.Response;
+
+public interface  Condition {
+
+    void  check(Response response);
+}
