+++ api-test/src/main/java/com/cinema/api/conditions/StatusCodeCondition.java
@@ -1,0 +1,15 @@
+package com.cinema.api.conditions;
+
+import io.restassured.response.Response;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class StatusCodeCondition implements Condition{
+
+    private final int statusCode;
+    @Override
+    public void check(Response response) {
+        response.then().assertThat().statusCode(statusCode);
+
+    }
+}
