+++ api-test/src/main/java/com/cinema/api/services/ApiService.java
@@ -1,0 +1,14 @@
+package com.cinema.api.services;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import io.restassured.specification.RequestSpecification;
+
+public  class ApiService {
+
+    protected RequestSpecification setup() {
+        return RestAssured
+                .given().contentType(ContentType.JSON).log().all();
+
+    }
+}
