--- api-test/src/main/java/com/socks/api/ProjectConfig.java
@@ -1,5 +1,0 @@
-package com.socks.api;
-
-public interface ProjectConfig {
-    //можно сделать юрл ссылку и ссылаться на неё
-}
