--- api-test/src/main/java/com/socks/api/conditions/BodyFieldCondition.java
@@ -1,17 +1,0 @@
-package com.socks.api.conditions;
-
-import io.restassured.response.Response;
-import lombok.RequiredArgsConstructor;
-import org.hamcrest.Matcher;
-
-@RequiredArgsConstructor
-public class BodyFieldCondition implements Condition {
-
-    private final String jsonPath;
-    private final Matcher matcher;
-
-    @Override
-    public void check(Response response) {
-        response.then().assertThat().body(jsonPath, matcher);
-    }
-}
