--- api-test/src/main/java/com/socks/api/conditions/Condition.java
@@ -1,8 +1,0 @@
-package com.socks.api.conditions;
-
-import io.restassured.response.Response;
-
-public interface  Condition {
-
-    void  check(Response response);
-}
