--- api-test/src/main/java/com/socks/api/conditions/StatusCodeCondition.java
@@ -1,15 +1,0 @@
-package com.socks.api.conditions;
-
-import io.restassured.response.Response;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class StatusCodeCondition implements Condition{
-
-    private final int statusCode;
-    @Override
-    public void check(Response response) {
-        response.then().assertThat().statusCode(statusCode);
-
-    }
-}
