--- api-test/src/main/java/com/socks/api/services/ApiService.java
@@ -1,14 +1,0 @@
-package com.socks.api.services;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import io.restassured.specification.RequestSpecification;
-
-public  class ApiService {
-
-    protected RequestSpecification setup() {
-        return RestAssured
-                .given().contentType(ContentType.JSON).log().all();
-
-    }
-}
