+++ ui-test/src/main/java/com/cinema/ui/LoggedUserPage.java
