+++ ui-test/src/main/java/com/cinema/ui/PaymentConfig.java
@@ -1,0 +1,21 @@
+package com.cinema.ui;
+
+import org.aeonbits.owner.Config;
+
+@Config.Sources("classpath:payment.properties")
+public interface PaymentConfig extends Config {
+    @Key("card.number")
+    String cardNumber();
+
+    @Key("card.cardholderName")
+    String cardholderName();
+
+    @Key("card.expiryMonth")
+    String month();
+
+    @Key("card.expiryYear")
+    String year();
+
+    @Key("card.cvc")
+    String cvc();
+}
