+++ ui-test/src/main/resources/payment.properties
@@ -1,0 +1,5 @@
+card.number=4242424242424242
+card.cardholderName=Bober
+card.expiryMonth=???
+card.expiryYear=2025
+card.cvc=123
