+++ ui-test/src/test/java/com/cinema/ui/test/BaseUiTest.java
@@ -1,0 +1,20 @@
+package com.cinema.ui.test;
+
+import com.codeborne.selenide.Selenide;
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+
+
+import static com.codeborne.selenide.Selenide.$;
+import static com.codeborne.selenide.Selenide.open;
+
+public class BaseUiTest {
+    //Конфиги типа Base URL
+    @BeforeEach
+    public void setUp() {
+    }
+
+    protected <T> T at(Class<T> pageClass) {
+        return Selenide.page(pageClass);
+    }
+}
