--- ui-test/src/test/java/com/socks/ui/test/BuyTicketTest.java
@@ -1,17 +1,0 @@
-package com.socks.ui.test;
-
-import com.socks.api.conditions.Condition;
-import com.socks.ui.MoviesPage;
-import com.socks.ui.MoviesPayPage;
-import org.junit.jupiter.api.Test;
-
-import static com.codeborne.selenide.Condition.text;
-
-public class BuyTicketTest extends  BaseUiTest {
-    @Test
-    public void userCantBuyTicketWithWrongCredentials() {
-        MoviesPage.open()
-            .openMovie();
-            at(MoviesPayPage.class);
-    }
-}
